dom/indexedDB/test/test_filehandle_store_snapshot.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript">
  function* testSteps()
  {
    const READ_WRITE = "readwrite";

    const name = window.location.pathname;

    const objectStoreName = "Blobs";

    const testFile = getRandomFile("random.bin", 100000);

    let request = indexedDB.open(name, 1);
    request.onerror = errorHandler;
    request.onupgradeneeded = grabEventAndContinueHandler;
    request.onsuccess = grabEventAndContinueHandler;
    let event = yield undefined;

    is(event.type, "upgradeneeded", "Got correct event type");

    let db = event.target.result;
    db.onerror = errorHandler;

    let objectStore = db.createObjectStore(objectStoreName, { });

    event = yield undefined;

    is(event.type, "success", "Got correct event type");

    request = db.createMutableFile("random.bin", "binary/random");
    request.onsuccess = grabEventAndContinueHandler;
    event = yield undefined;

    let mutableFile = event.target.result;
    mutableFile.onerror = errorHandler;

    let fileHandle = mutableFile.open("readwrite");

    is(getFileId(mutableFile), 1, "Correct file id");

    request = fileHandle.write(testFile);
    request.onsuccess = grabEventAndContinueHandler;
    event = yield undefined;

    request = mutableFile.getFile();
    request.onsuccess = grabEventAndContinueHandler;
    event = yield undefined;

    let file = event.target.result;

    let trans = db.transaction([objectStoreName], READ_WRITE);
    objectStore = trans.objectStore(objectStoreName);

    request = objectStore.add(file, 42);
    request.onsuccess = grabEventAndContinueHandler;
    event = yield undefined;

    // At this moment, the file should not be readable anymore.
    let reader = new FileReader();
    try {
      reader.readAsArrayBuffer(file);
      ok(false, "Should have thrown!");
    }
    catch (e) {
      ok(e instanceof DOMException, "Got exception.");
      is(e.name, "FileHandleInactiveError", "Good error.");
      is(e.code, 0, "Good error code.");
    }

    request = objectStore.get(42);
    request.onsuccess = grabEventAndContinueHandler;
    event = yield undefined;

    verifyBlob(event.target.result, testFile, 2);
    yield undefined;

    finishTest();
  }
  </script>
  <script type="text/javascript" src="file.js"></script>
  <script type="text/javascript" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
